fix CD deployment failure case
