+++ src/pkg/cli/client/byoc/gcp/byoc.go
@@ -18,6 +18,7 @@ import (
 	"cloud.google.com/go/storage"
 	"github.com/DefangLabs/defang/src/pkg"
 	"github.com/DefangLabs/defang/src/pkg/cli/client"
+	cliClient "github.com/DefangLabs/defang/src/pkg/cli/client"
 	"github.com/DefangLabs/defang/src/pkg/cli/client/byoc"
 	"github.com/DefangLabs/defang/src/pkg/cli/compose"
 	"github.com/DefangLabs/defang/src/pkg/clouds/aws/ecs"
@@ -411,26 +412,26 @@ func (b *ByocGcp) GetDeploymentStatus(ctx context.Context) error {
 	}
 
 	var completionTime = execution.GetCompletionTime()
-	if completionTime == nil {
-		// still running
-		return nil
-	} else {
+	if completionTime != nil {
 		// cd is done
-		var msgs []string
 		var failedTasks = execution.GetFailedCount()
 		if failedTasks > 0 {
+			var msgs []string
 			for _, condition := range execution.GetConditions() {
 				if condition.GetType() == "Completed" && condition.GetMessage() != "" {
 					msgs = append(msgs, condition.GetMessage())
 				}
 			}
 
-			return pkg.ErrDeploymentFailed{Message: strings.Join(msgs, ",")}
+			return cliClient.ErrDeploymentFailed{Message: strings.Join(msgs, ",")}
 		}
 
 		// completed successfully
 		return io.EOF
 	}
+
+	// still running
+	return nil
 }
 
 func (b *ByocGcp) deploy(ctx context.Context, req *defangv1.DeployRequest, command string) (*defangv1.DeployResponse, error) {
